requirements.txt
requests==2.27.1
asgiref==3.2.10
Django==3.1
django-admin-sortable2==0.7.8
django-tinymce==3.4.0
environs==9.5.0
marshmallow==3.15.0
packaging==21.3
Pillow==8.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
